requirements.txt
alabaster==1.0.0
amqp==5.3.1
annotated-types==0.7.0
ansicon==1.89.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
billiard==4.2.1
blessed==1.21.0
botocore==1.39.14
bpython==0.25
celery==5.5.3
certifi==2024.8.30
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.5
curio==1.6
curtsies==0.4.3
cwcwidth==0.1.10
cycler==0.12.1
dask==2025.7.0
distro==1.9.0
Django==5.1.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.5.0
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
docutils==0.21.2
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fanstatic==1.5
fastapi==0.116.1
fastapi-imp==0.0.1
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.7.0
greenlet==3.2.3
h11==0.14.0
hid_converge==2.0.1.1
HTMLParser==0.0.2
httpcore==1.0.8
httpsproxy_urllib2==1.0
httpx==0.28.1
hypothesis==6.136.4
idna==3.10
imagesize==1.4.1
import-java==0.6
importlib_metadata==8.7.0
iniconfig==2.1.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
Jinja2==3.1.6
jinxed==1.3.0
jiter==0.7.0
jmespath==1.0.1
js==1.0
keyring==25.6.0
kiwisolver==1.4.8
kombu==5.5.4
language_tool_python==2.8.1
locket==1.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
olefile==0.47
openai==1.54.3
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
partd==1.4.2
pillow==11.0.0
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg2-binary==2.9.10
PyConfigParser==1.0.5
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.19.2
pyjnius==1.6.1
PyJWT==2.10.1
pymemcache==4.0.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
PySocks==1.7.1
PySoundFile==0.9.0.post1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-magic==0.4.27
python-ntlm==1.1.0
pytz==2025.2
pywin32-ctypes==0.2.3
pyxdg==0.28
PyYAML==6.0.2
redis==6.2.0
requests==2.32.4
rich==14.1.0
roman-numerals-py==3.1.0
selenium==4.32.0
setuptools==80.9.0
shiboken6==6.9.1
simplejson==3.20.1
six==1.16.0
sniffio==1.3.1
snowballstemmer==3.0.1
sortedcontainers==2.4.0
soupsieve==2.7
SpeechRecognition==3.11.0
Sphinx==8.3.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.4.4
srt==3.5.3
starlette==0.47.2
sympy==1.14.0
toolz==1.0.0
torch==2.7.1
tqdm==4.67.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
vosk==0.3.45
wcwidth==0.2.13
WebOb==1.8.9
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
whitenoise==6.8.2
wsproto==1.2.0
xlsxwriter==3.2.5
zipp==3.23.0
